requirements.txt
pandas==2.2.2
Pillow==10.4.0
plotly==5.9.0
st_social_media_links==0.1.3
streamlit==1.36.0
